\name{Bli}
\alias{Bli}
\title{
Constructor of "Bli" class
}
\description{
Construction of a new "Bli" object with all slots initialized.
}
\usage{
Bli(...)
}
\arguments{
  \item{\dots}{
.Object in \code{initialize}; leave empty
}
}
\value{
S4 object of "Bli" class
}
\references{
\insertRef{pza}{smoke}
}
\author{
\packageAuthor{smoke}

Maintainer: \packageMaintainer{smoke}
}

\seealso{
 \code{\link{Bli-class}}
}
\examples{
bli1 <- Bli()
}
\keyword{ manip }
